services/filecloud_cold/requirements.txt
boto3==1.17.112
psycopg2-binary==2.9.1
python-dotenv==0.18.0
s3transfer==0.4.2
jsonref==0.2
jsonschema==3.2.0
PyYAML==5.4.1
python-dotenv==0.18.0